SIGQUITのシグナルハンドラ復元
